backend_proper/requirements.txt
fastapi
uvicorn
python-dotenv
motor
bcrypt
python-jose[cryptography]
